integrations/requirements.txt
slack-bolt>=1.18.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
discord.py>=2.3.0